deepseek_ocr_desktop/pyproject.toml
[project]
name = "deepseek-ocr-desktop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.34.2",
    "addict>=2.4.0",
    "easydict>=1.13",
    "einops>=0.8.1",
    "img2pdf>=0.5.0",
    "markdown>=3.5.0",
    "pillow>=10.0.0",
    "pyinstaller>=6.0.0",
    "pymupdf>=1.23.0",
    "pyside6>=6.6.0",
    "python-docx>=1.1.0",
    "safetensors>=0.7.0",
    "tokenizers==0.20.3",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers==4.46.3",
]
